+++ src/main/java/ladder/wrapper/LadderNumber.java
@@ -1,6 +1,4 @@
-package ladder;
-
-import static ladder.Direction.*;
+package ladder.wrapper;
 
 public class LadderNumber {
