+++ src/main/java/ladder/wrapper/NumberOfPerson.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.wrapper;
 
 public class NumberOfPerson {
