+++ src/main/java/ladder/wrapper/NumberOfRow.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.wrapper;
 
 public class NumberOfRow {
