+++ src/test/java/ladder/LadderGamePositionFactoryTest.java
@@ -1,5 +1,6 @@
 package ladder;
 
+import ladder.wrapper.Position;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
